fix(build): alias, esbuild, treeshake, single file
