+++ package.json
@@ -5,7 +5,7 @@
   "main": "dist/server.js",
   "scripts": {
     "dev": "nodemon src/server.ts",
-    "build": "rimraf dist && rollup -c rollup.config.mjs",
+    "build": "rimraf dist && rollup -c rollup.config.js",
     "start": "node -r dotenv/config dist/server.js",
     "lint": "next lint",
     "test": "vitest",
@@ -14,28 +14,33 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "type": "module",
   "devDependencies": {
+    "@rollup/plugin-alias": "^5.1.0",
+    "@rollup/plugin-commonjs": "^25.0.7",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-typescript": "^11.1.6",
+    "@sentry/rollup-plugin": "^2.14.2",
     "@types/express": "^4.17.21",
     "@types/node": "^20.11.19",
+    "esbuild": "^0.20.1",
     "eslint": "^8.56.0",
     "nodemon": "^3.0.3",
     "rimraf": "^5.0.5",
     "rollup": "^4.12.0",
+    "rollup-plugin-auto-external": "^2.0.0",
+    "rollup-plugin-esbuild": "^6.1.1",
+    "rollup-plugin-polyfill-node": "^0.13.0",
     "tsx": "^4.7.1",
     "typescript": "^5.3.3",
     "typescript-eslint": "^7.0.2",
     "vitest": "^1.3.1"
   },
   "dependencies": {
-    "@hypercerts-org/sdk": "^1.4.1",
-    "@rollup/plugin-commonjs": "^25.0.7",
+    "@hypercerts-org/sdk": "^1.4.2-alpha.2",
     "@sentry/integrations": "^7.102.1",
     "@sentry/node": "^7.102.1",
     "@sentry/profiling-node": "^7.102.1",
-    "@sentry/rollup-plugin": "^2.14.2",
     "@supabase/supabase-js": "^2.39.7",
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
